+++ lib/src/entities/micro_app_preferences.dart
@@ -1,8 +1,12 @@
-import '../utils/constants/path_separator.dart';
+import 'package:flutter_micro_app/flutter_micro_app.dart';
+
+final maAppBaseRoute = MicroAppPreferences.config.appBaseRoute.baseRoute.route;
 
 class MicroAppPreferences {
   static MicroAppConfig config = MicroAppConfig(
-      nativeEventsEnabled: false, pathSeparator: MicroAppPathSeparator.dot);
+      nativeEventsEnabled: false,
+      pathSeparator: MicroAppPathSeparator.slash,
+      appBaseRoute: _DefaultBaseRoute());
   static update(MicroAppConfig _config) {
     config = _config;
   }
@@ -17,11 +21,40 @@ class MicroAppConfig {
   /// pathSeparator route segments
   final String pathSeparator;
 
-  MicroAppConfig(
-      {required this.nativeEventsEnabled, required this.pathSeparator});
+  /// Default baseRoute = "/"
+  final MicroAppBaseRoute appBaseRoute;
 
-  MicroAppConfig copyWith({bool? nativeEventsEnabled, String? pathSeparator}) =>
+  /// disable all console logs
+  final bool consoleLogsEnabled;
+
+  /// Default page transition
+  final MicroPageTransitionType pageTransitionType;
+
+  MicroAppConfig(
+      {this.nativeEventsEnabled = false,
+      this.pathSeparator = MicroAppPathSeparator.slash,
+      this.consoleLogsEnabled = false,
+      MicroAppBaseRoute? appBaseRoute,
+      this.pageTransitionType = MicroPageTransitionType.platform})
+      : appBaseRoute = appBaseRoute ?? _DefaultBaseRoute();
+
+  /// Protype [MicroAppConfig]
+  MicroAppConfig copyWith({
+    bool? nativeEventsEnabled,
+    String? pathSeparator,
+    MicroAppBaseRoute? appBaseRoute,
+    bool? consoleLogsEnabled,
+    MicroPageTransitionType? pageTransitionType,
+  }) =>
       MicroAppConfig(
           nativeEventsEnabled: nativeEventsEnabled ?? this.nativeEventsEnabled,
-          pathSeparator: pathSeparator ?? this.pathSeparator);
+          pathSeparator: pathSeparator ?? this.pathSeparator,
+          appBaseRoute: appBaseRoute ?? _DefaultBaseRoute(),
+          consoleLogsEnabled: consoleLogsEnabled ?? this.consoleLogsEnabled,
+          pageTransitionType: pageTransitionType ?? this.pageTransitionType);
+}
+
+class _DefaultBaseRoute extends MicroAppBaseRoute {
+  @override
+  MicroAppRoute get baseRoute => MicroAppRoute('/');
 }